+++ CHANGES.md
@@ -7,6 +7,7 @@ v1.2 (YYYY-MM-DD)
 
 - Added `-v` option to `stringsutil report` sub-command to show unlocalized
   strings.
+- Added Windows API support to the `sfSetLocale` function.
 - Updated the output from `stringsutil translate` to better show progress.
 - Updated code to work with latest CUPS 2.5/3.0.
 - Fixed a crash bug in `stringsutil report` when checking format strings.